+++ lib/screens/splash.dart
@@ -1,0 +1,36 @@
+import 'dart:async';
+
+import 'package:flutter/material.dart';
+
+import 'home.dart';
+
+class SplashScreen extends StatefulWidget {
+  @override
+  _SplashScreenState createState() => new _SplashScreenState();
+}
+
+class _SplashScreenState extends State<SplashScreen> {
+  @override
+  void initState() {
+    super.initState();
+    Timer(
+      Duration(seconds: 3),
+      () => Navigator.pushReplacement(
+        context,
+        MaterialPageRoute(
+          builder: (context) => HomePage(),
+        ),
+      ),
+    );
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return new Scaffold(
+      backgroundColor: Colors.black,
+      body: new Center(
+        child: new Image.asset('assets/images/logo.png'),
+      ),
+    );
+  }
+}
